feat: added sample experiment
- added to tmp/experiments/sample_experiment.rb when running `rails phlex_storybook:install`
