+++ lib/phlex_storybook/components/experiments/index.rb
@@ -13,7 +13,7 @@ module PhlexStorybook
           turbo_frame_tag("story_components") do
             div class: "flex h-screen w-screen" do
               div class: "flex-none w-1/4 min-w-40 max-w-80 bg-slate-700 text-white story-selector" do
-                render Stories::ComponentSelector.new(story_components: @story_components, selected: @selected)
+                render SidebarMenu.new(story_components: @story_components, selected: @selected)
               end
 
               div class: "flex-auto h-full w-3/4 bg-white" do
